+++ readium/shared/CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## Added
+
+* `Publication.Service.Context` now holds a reference to the parent `Publication`. This can be used to access other services from a given `Publication.Service` implementation.
+* The default `LocatorService` implementation can be used to get a `Locator` from a global progression in the publication.
+  * `publication.locateProgression(0.5)`
+
+
 ## [2.0.0-beta.1]
 
 ### Added